backend/requirements.txt
# --- Web Framework ---
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9

# --- Video / Image Processing ---
imageio==2.34.0
imageio-ffmpeg==0.4.9
opencv-python==4.10.0.84
ffmpeg-python==0.2.0

# --- AI / Diffusion Models ---
torch>=2.0.0
diffusers>=0.28.0
transformers>=4.41.0
accelerate>=0.30.0
safetensors>=0.4.3

# --- Utils ---
numpy>=1.26.4
pillow>=10.3.0
